Possible improved version of rrmxmx by Pelle Evensen
The testing code is in mult's LinnormA, and is very similar to rrmxmx except that it uses a xor-rotate-xor-rotate in place of where he had used a xorshift, and any multiplications are instead XLCG "xor by something ending in 0b101, multiply by something ending in 0b011" operations. The multiplications use much smaller constants, which could help speed a little.