Set up a crude replica of Pelle Evensen's rrxmrrxmsx test harness
I'm testing rrxmrrxmsx_0 with different multiplicative constants, which has currently passed the reversed versions of left-rotates 0 to 19 (or right rotates 45 to 63 and 0), as well as some others, up to 1TB. The tests I'm doing use `-tf 2` but not `-te 1`, because I didn't know that `-te 1` was supposed to be used and when I started some simple tests with `-te 1` on another computer it took much longer. The new multipliers seem good so far. 0xAEF17502108EF2D9 is used in PCG-Random as a multiplier after a variable shift or rotate. 0xDB4F0B9175AE2165 is related to a harmonious number (generalized golden ratio); it is 2 to the 64 divided by the fourth harmonious number, with the last bit set to 1 if necessary. Neither is prime.