Sashimi32 is a new 32-bit Xoshiro scrambler
It passes 32TB of PractRand with one anomaly (it probably can pass with no anomalies on some seeds, due to the large state size of 2 to the 128 minus 1). Speed is unclear at this point, but it should resist the easy ways to accidentally weaken XoshiroAra/Xara/StarPhi and the sorta-less-frequent ways to weaken XoshiroStarStar.