Add TwinLinear, try various other changes
The new variant on MizuchiRNG seems to do well with selectable streams. Ladder32RNG uses less multiplication than Mover32RNG but isn't as fast usually. TwinLinear, from a paper by anonymous authors (but probably involving Guy Steele Jr., let's not kid ourselves here) is implemented and tweaked a bit for speed. Unforunately, in no form will TwinLinear be equidistributed, it is unlikely to be very fast, and a state size of 128 bits plus an optional stream size of 126 bits is ridiculous for a period of 2 to the 64 (even  if you consider that per-stream). The tweaks improve period slightly, but I don't think they help much (maybe 2 to the 68 instead of 2 to the 64).