Half-way test Romu, fully test Orbit (now fast!)
Romu will probably pass PractRand becauase of how much state it has, but the rate of early anomalies it has is somewhat concerning for what is supposed to be a high-quality generator. Orbit (tested in mingler) gets one anomaly and passes 32TB, and its period is guaranteed longer than RomoDuo at 2 to the 128.