Add ACORN, order 10; as expected, it fails immediately.
ACORN can only be expected to pass for a decent generator if the low-order bits are ignored, as with a generator that produces only floats and tests that only work with floats. Even then, a good test like PractRand can pick up easy patterns in the output, and the period is incredibly bad for how much state it uses (this one has a period of ((2 to the 64) * 10), with 640 bits of state and 64 bits for a stream).